+++ app/src/main/java/com/example/simple_live_data_sample/ViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.calclivedataandbind
+package com.example.simple_live_data_sample
 
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
